requirements.txt
pyrogram==2.0.37
tgcrypto==1.2.3
httpx==0.24.1
Jinja2==3.1.2
Flask==2.3.2
python-dotenv==1.0.0
